+++ swarm/src/behaviour.rs
@@ -24,23 +24,22 @@ mod listen_addresses;
 mod peer_addresses;
 pub mod toggle;
 
-pub use external_addresses::ExternalAddresses;
-pub use listen_addresses::ListenAddresses;
-pub use peer_addresses::PeerAddresses;
+use std::task::{Context, Poll};
 
-use crate::connection::ConnectionId;
-use crate::dial_opts::DialOpts;
-use crate::listen_opts::ListenOpts;
-use crate::{
-    ConnectionDenied, ConnectionError, ConnectionHandler, DialError, ListenError, THandler,
-    THandlerInEvent, THandlerOutEvent,
-};
+pub use external_addresses::ExternalAddresses;
 use libp2p_core::{
     transport::{ListenerId, PortUse},
     ConnectedPoint, Endpoint, Multiaddr,
 };
 use libp2p_identity::PeerId;
-use std::{task::Context, task::Poll};
+pub use listen_addresses::ListenAddresses;
+pub use peer_addresses::PeerAddresses;
+
+use crate::{
+    connection::ConnectionId, dial_opts::DialOpts, listen_opts::ListenOpts, ConnectionDenied,
+    ConnectionError, ConnectionHandler, DialError, ListenError, THandler, THandlerInEvent,
+    THandlerOutEvent,
+};
 
 /// A [`NetworkBehaviour`] defines the behaviour of the local node on the network.
 ///
@@ -101,25 +100,25 @@ use std::{task::Context, task::Poll};
 /// #[behaviour(to_swarm = "Event")]
 /// # #[behaviour(prelude = "libp2p_swarm::derive_prelude")]
 /// struct MyBehaviour {
-///   identify: identify::Behaviour,
-///   ping: ping::Behaviour,
+///     identify: identify::Behaviour,
+///     ping: ping::Behaviour,
 /// }
 ///
 /// enum Event {
-///   Identify(identify::Event),
-///   Ping(ping::Event),
+///     Identify(identify::Event),
+///     Ping(ping::Event),
 /// }
 ///
 /// impl From<identify::Event> for Event {
-///   fn from(event: identify::Event) -> Self {
-///     Self::Identify(event)
-///   }
+///     fn from(event: identify::Event) -> Self {
+///         Self::Identify(event)
+///     }
 /// }
 ///
 /// impl From<ping::Event> for Event {
-///   fn from(event: ping::Event) -> Self {
-///     Self::Ping(event)
-///   }
+///     fn from(event: ping::Event) -> Self {
+///         Self::Ping(event)
+///     }
 /// }
 /// ```
 pub trait NetworkBehaviour: 'static {
@@ -131,8 +130,8 @@ pub trait NetworkBehaviour: 'static {
 
     /// Callback that is invoked for every new inbound connection.
     ///
-    /// At this point in the connection lifecycle, only the remote's and our local address are known.
-    /// We have also already allocated a [`ConnectionId`].
+    /// At this point in the connection lifecycle, only the remote's and our local address are
+    /// known. We have also already allocated a [`ConnectionId`].
     ///
     /// Any error returned from this function will immediately abort the dial attempt.
     fn handle_pending_inbound_connection(
@@ -148,9 +147,10 @@ pub trait NetworkBehaviour: 'static {
     ///
     /// This is invoked once another peer has successfully dialed us.
     ///
-    /// At this point, we have verified their [`PeerId`] and we know, which particular [`Multiaddr`] succeeded in the dial.
-    /// In order to actually use this connection, this function must return a [`ConnectionHandler`].
-    /// Returning an error will immediately close the connection.
+    /// At this point, we have verified their [`PeerId`] and we know, which particular [`Multiaddr`]
+    /// succeeded in the dial. In order to actually use this connection, this function must
+    /// return a [`ConnectionHandler`]. Returning an error will immediately close the
+    /// connection.
     ///
     /// Note when any composed behaviour returns an error the connection will be closed and a
     /// [`FromSwarm::ListenFailure`] event will be emitted.
@@ -168,10 +168,14 @@ pub trait NetworkBehaviour: 'static {
     ///
     /// - The [`PeerId`], if known. Remember that we can dial without a [`PeerId`].
     /// - All addresses passed to [`DialOpts`] are passed in here too.
-    /// - The effective [`Role`](Endpoint) of this peer in the dial attempt. Typically, this is set to [`Endpoint::Dialer`] except if we are attempting a hole-punch.
-    /// - The [`ConnectionId`] identifying the future connection resulting from this dial, if successful.
+    /// - The effective [`Role`](Endpoint) of this peer in the dial attempt. Typically, this is set
+    ///   to [`Endpoint::Dialer`] except if we are attempting a hole-punch.
+    /// - The [`ConnectionId`] identifying the future connection resulting from this dial, if
+    ///   successful.
     ///
-    /// Note that the addresses returned from this function are only used for dialing if [`WithPeerIdWithAddresses::extend_addresses_through_behaviour`](crate::dial_opts::WithPeerIdWithAddresses::extend_addresses_through_behaviour) is set.
+    /// Note that the addresses returned from this function are only used for dialing if
+    /// [`WithPeerIdWithAddresses::extend_addresses_through_behaviour`](crate::dial_opts::WithPeerIdWithAddresses::extend_addresses_through_behaviour)
+    /// is set.
     ///
     /// Any error returned from this function will immediately abort the dial attempt.
     fn handle_pending_outbound_connection(
@@ -187,9 +191,10 @@ pub trait NetworkBehaviour: 'static {
     /// Callback that is invoked for every established outbound connection.
     ///
     /// This is invoked once we have successfully dialed a peer.
-    /// At this point, we have verified their [`PeerId`] and we know, which particular [`Multiaddr`] succeeded in the dial.
-    /// In order to actually use this connection, this function must return a [`ConnectionHandler`].
-    /// Returning an error will immediately close the connection.
+    /// At this point, we have verified their [`PeerId`] and we know, which particular [`Multiaddr`]
+    /// succeeded in the dial. In order to actually use this connection, this function must
+    /// return a [`ConnectionHandler`]. Returning an error will immediately close the
+    /// connection.
     ///
     /// Note when any composed behaviour returns an error the connection will be closed and a
     /// [`FromSwarm::DialFailure`] event will be emitted.
@@ -240,8 +245,9 @@ pub enum ToSwarm<TOutEvent, TInEvent> {
     /// On failure, [`NetworkBehaviour::on_swarm_event`] with `DialFailure` is invoked.
     ///
     /// [`DialOpts`] provides access to the [`ConnectionId`] via [`DialOpts::connection_id`].
-    /// This [`ConnectionId`] will be used throughout the connection's lifecycle to associate events with it.
-    /// This allows a [`NetworkBehaviour`] to identify a connection that resulted out of its own dial request.
+    /// This [`ConnectionId`] will be used throughout the connection's lifecycle to associate
+    /// events with it. This allows a [`NetworkBehaviour`] to identify a connection that
+    /// resulted out of its own dial request.
     Dial { opts: DialOpts },
 
     /// Instructs the [`Swarm`](crate::Swarm) to listen on the provided address.
@@ -253,8 +259,8 @@ pub enum ToSwarm<TOutEvent, TInEvent> {
     /// Instructs the `Swarm` to send an event to the handler dedicated to a
     /// connection with a peer.
     ///
-    /// If the `Swarm` is connected to the peer, the message is delivered to the [`ConnectionHandler`]
-    /// instance identified by the peer ID and connection ID.
+    /// If the `Swarm` is connected to the peer, the message is delivered to the
+    /// [`ConnectionHandler`] instance identified by the peer ID and connection ID.
     ///
     /// If the specified connection no longer exists, the event is silently dropped.
     ///
@@ -278,11 +284,12 @@ pub enum ToSwarm<TOutEvent, TInEvent> {
     ///
     /// The emphasis on a **new** candidate is important.
     /// Protocols MUST take care to only emit a candidate once per "source".
-    /// For example, the observed address of a TCP connection does not change throughout its lifetime.
-    /// Thus, only one candidate should be emitted per connection.
+    /// For example, the observed address of a TCP connection does not change throughout its
+    /// lifetime. Thus, only one candidate should be emitted per connection.
     ///
-    /// This makes the report frequency of an address a meaningful data-point for consumers of this event.
-    /// This address will be shared with all [`NetworkBehaviour`]s via [`FromSwarm::NewExternalAddrCandidate`].
+    /// This makes the report frequency of an address a meaningful data-point for consumers of this
+    /// event. This address will be shared with all [`NetworkBehaviour`]s via
+    /// [`FromSwarm::NewExternalAddrCandidate`].
     ///
     /// This address could come from a variety of sources:
     /// - A protocol such as identify obtained it from a remote.
@@ -290,25 +297,32 @@ pub enum ToSwarm<TOutEvent, TInEvent> {
     /// - We made an educated guess based on one of our listen addresses.
     NewExternalAddrCandidate(Multiaddr),
 
-    /// Indicates to the [`Swarm`](crate::Swarm) that the provided address is confirmed to be externally reachable.
+    /// Indicates to the [`Swarm`](crate::Swarm) that the provided address is confirmed to be
+    /// externally reachable.
     ///
-    /// This is intended to be issued in response to a [`FromSwarm::NewExternalAddrCandidate`] if we are indeed externally reachable on this address.
-    /// This address will be shared with all [`NetworkBehaviour`]s via [`FromSwarm::ExternalAddrConfirmed`].
+    /// This is intended to be issued in response to a [`FromSwarm::NewExternalAddrCandidate`] if
+    /// we are indeed externally reachable on this address. This address will be shared with
+    /// all [`NetworkBehaviour`]s via [`FromSwarm::ExternalAddrConfirmed`].
     ExternalAddrConfirmed(Multiaddr),
 
-    /// Indicates to the [`Swarm`](crate::Swarm) that we are no longer externally reachable under the provided address.
+    /// Indicates to the [`Swarm`](crate::Swarm) that we are no longer externally reachable under
+    /// the provided address.
     ///
     /// This expires an address that was earlier confirmed via [`ToSwarm::ExternalAddrConfirmed`].
-    /// This address will be shared with all [`NetworkBehaviour`]s via [`FromSwarm::ExternalAddrExpired`].
+    /// This address will be shared with all [`NetworkBehaviour`]s via
+    /// [`FromSwarm::ExternalAddrExpired`].
     ExternalAddrExpired(Multiaddr),
 
-    /// Instructs the `Swarm` to initiate a graceful close of one or all connections with the given peer.
+    /// Instructs the `Swarm` to initiate a graceful close of one or all connections with the given
+    /// peer.
     ///
-    /// Closing a connection via [`ToSwarm::CloseConnection`] will poll [`ConnectionHandler::poll_close`] to completion.
-    /// In most cases, stopping to "use" a connection is enough to have it closed.
-    /// The keep-alive algorithm will close a connection automatically once all [`ConnectionHandler`]s are idle.
+    /// Closing a connection via [`ToSwarm::CloseConnection`] will poll
+    /// [`ConnectionHandler::poll_close`] to completion. In most cases, stopping to "use" a
+    /// connection is enough to have it closed. The keep-alive algorithm will close a
+    /// connection automatically once all [`ConnectionHandler`]s are idle.
     ///
-    /// Use this command if you want to close a connection _despite_ it still being in use by one or more handlers.
+    /// Use this command if you want to close a connection _despite_ it still being in use by one
+    /// or more handlers.
     CloseConnection {
         /// The peer to disconnect.
         peer_id: PeerId,
@@ -316,7 +330,8 @@ pub enum ToSwarm<TOutEvent, TInEvent> {
         connection: CloseConnection,
     },
 
-    /// Reports external address of a remote peer to the [`Swarm`](crate::Swarm) and through that to other [`NetworkBehaviour`]s.
+    /// Reports external address of a remote peer to the [`Swarm`](crate::Swarm) and through that
+    /// to other [`NetworkBehaviour`]s.
     NewExternalAddrOfPeer { peer_id: PeerId, address: Multiaddr },
 }
 
@@ -440,8 +455,8 @@ pub enum FromSwarm<'a> {
     /// Informs the behaviour that an error
     /// happened on an incoming connection during its initial handshake.
     ///
-    /// This can include, for example, an error during the handshake of the encryption layer, or the
-    /// connection unexpectedly closed.
+    /// This can include, for example, an error during the handshake of the encryption layer, or
+    /// the connection unexpectedly closed.
     ListenFailure(ListenFailure<'a>),
     /// Informs the behaviour that a new listener was created.
     NewListener(NewListener),
@@ -455,11 +470,13 @@ pub enum FromSwarm<'a> {
     ListenerError(ListenerError<'a>),
     /// Informs the behaviour that a listener closed.
     ListenerClosed(ListenerClosed<'a>),
-    /// Informs the behaviour that we have discovered a new candidate for an external address for us.
+    /// Informs the behaviour that we have discovered a new candidate for an external address for
+    /// us.
     NewExternalAddrCandidate(NewExternalAddrCandidate<'a>),
     /// Informs the behaviour that an external address of the local node was confirmed.
     ExternalAddrConfirmed(ExternalAddrConfirmed<'a>),
-    /// Informs the behaviour that an external address of the local node expired, i.e. is no-longer confirmed.
+    /// Informs the behaviour that an external address of the local node expired, i.e. is no-longer
+    /// confirmed.
     ExternalAddrExpired(ExternalAddrExpired<'a>),
     /// Informs the behaviour that we have discovered a new external address for a remote peer.
     NewExternalAddrOfPeer(NewExternalAddrOfPeer<'a>),
@@ -559,7 +576,8 @@ pub struct ListenerClosed<'a> {
     pub reason: Result<(), &'a std::io::Error>,
 }
 
-/// [`FromSwarm`] variant that informs the behaviour about a new candidate for an external address for us.
+/// [`FromSwarm`] variant that informs the behaviour about a new candidate for an external address
+/// for us.
 #[derive(Debug, Clone, Copy)]
 pub struct NewExternalAddrCandidate<'a> {
     pub addr: &'a Multiaddr,
@@ -577,7 +595,8 @@ pub struct ExternalAddrExpired<'a> {
     pub addr: &'a Multiaddr,
 }
 
-/// [`FromSwarm`] variant that informs the behaviour that a new external address for a remote peer was detected.
+/// [`FromSwarm`] variant that informs the behaviour that a new external address for a remote peer
+/// was detected.
 #[derive(Clone, Copy, Debug)]
 pub struct NewExternalAddrOfPeer<'a> {
     pub peer_id: PeerId,