+++ swarm/src/connection.rs
@@ -23,42 +23,47 @@ mod error;
 pub(crate) mod pool;
 mod supported_protocols;
 
+use std::{
+    collections::{HashMap, HashSet},
+    fmt,
+    fmt::{Display, Formatter},
+    future::Future,
+    io, mem,
+    pin::Pin,
+    sync::atomic::{AtomicUsize, Ordering},
+    task::{Context, Poll, Waker},
+    time::Duration,
+};
+
 pub use error::ConnectionError;
 pub(crate) use error::{
     PendingConnectionError, PendingInboundConnectionError, PendingOutboundConnectionError,
 };
-use libp2p_core::transport::PortUse;
-pub use supported_protocols::SupportedProtocols;
-
-use crate::handler::{
-    AddressChange, ConnectionEvent, ConnectionHandler, DialUpgradeError, FullyNegotiatedInbound,
-    FullyNegotiatedOutbound, ListenUpgradeError, ProtocolSupport, ProtocolsChange, UpgradeInfoSend,
+use futures::{future::BoxFuture, stream, stream::FuturesUnordered, FutureExt, StreamExt};
+use futures_timer::Delay;
+use libp2p_core::{
+    connection::ConnectedPoint,
+    multiaddr::Multiaddr,
+    muxing::{StreamMuxerBox, StreamMuxerEvent, StreamMuxerExt, SubstreamBox},
+    transport::PortUse,
+    upgrade,
+    upgrade::{NegotiationError, ProtocolError},
+    Endpoint,
 };
-use crate::stream::ActiveStreamCounter;
-use crate::upgrade::{InboundUpgradeSend, OutboundUpgradeSend};
+use libp2p_identity::PeerId;
+pub use supported_protocols::SupportedProtocols;
+use web_time::Instant;
+
 use crate::{
+    handler::{
+        AddressChange, ConnectionEvent, ConnectionHandler, DialUpgradeError,
+        FullyNegotiatedInbound, FullyNegotiatedOutbound, ListenUpgradeError, ProtocolSupport,
+        ProtocolsChange, UpgradeInfoSend,
+    },
+    stream::ActiveStreamCounter,
+    upgrade::{InboundUpgradeSend, OutboundUpgradeSend},
     ConnectionHandlerEvent, Stream, StreamProtocol, StreamUpgradeError, SubstreamProtocol,
 };
-use futures::future::BoxFuture;
-use futures::stream::FuturesUnordered;
-use futures::StreamExt;
-use futures::{stream, FutureExt};
-use futures_timer::Delay;
-use libp2p_core::connection::ConnectedPoint;
-use libp2p_core::multiaddr::Multiaddr;
-use libp2p_core::muxing::{StreamMuxerBox, StreamMuxerEvent, StreamMuxerExt, SubstreamBox};
-use libp2p_core::upgrade;
-use libp2p_core::upgrade::{NegotiationError, ProtocolError};
-use libp2p_core::Endpoint;
-use libp2p_identity::PeerId;
-use std::collections::{HashMap, HashSet};
-use std::fmt::{Display, Formatter};
-use std::future::Future;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::task::Waker;
-use std::time::Duration;
-use std::{fmt, io, mem, pin::Pin, task::Context, task::Poll};
-use web_time::Instant;
 
 static NEXT_CONNECTION_ID: AtomicUsize = AtomicUsize::new(1);
 
@@ -72,7 +77,8 @@ impl ConnectionId {
     /// [`Swarm`](crate::Swarm) enforces that [`ConnectionId`]s are unique and not reused.
     /// This constructor does not, hence the _unchecked_.
     ///
-    /// It is primarily meant for allowing manual tests of [`NetworkBehaviour`](crate::NetworkBehaviour)s.
+    /// It is primarily meant for allowing manual tests of
+    /// [`NetworkBehaviour`](crate::NetworkBehaviour)s.
     pub fn new_unchecked(id: usize) -> Self {
         Self(id)
     }
@@ -147,8 +153,8 @@ where
     max_negotiating_inbound_streams: usize,
     /// Contains all upgrades that are waiting for a new outbound substream.
     ///
-    /// The upgrade timeout is already ticking here so this may fail in case the remote is not quick
-    /// enough in providing us with a new stream.
+    /// The upgrade timeout is already ticking here so this may fail in case the remote is not
+    /// quick enough in providing us with a new stream.
     requested_substreams: FuturesUnordered<
         SubstreamRequested<THandler::OutboundOpenInfo, THandler::OutboundProtocol>,
     >,
@@ -223,7 +229,8 @@ where
         self.handler.on_behaviour_event(event);
     }
 
-    /// Begins an orderly shutdown of the connection, returning a stream of final events and a `Future` that resolves when connection shutdown is complete.
+    /// Begins an orderly shutdown of the connection, returning a stream of final events and a
+    /// `Future` that resolves when connection shutdown is complete.
     pub(crate) fn close(
         self,
     ) -> (
@@ -320,7 +327,8 @@ where
                 }
             }
 
-            // In case the [`ConnectionHandler`] can not make any more progress, poll the negotiating outbound streams.
+            // In case the [`ConnectionHandler`] can not make any more progress, poll the
+            // negotiating outbound streams.
             match negotiating_out.poll_next_unpin(cx) {
                 Poll::Pending | Poll::Ready(None) => {}
                 Poll::Ready(Some((info, Ok(protocol)))) => {
@@ -368,7 +376,8 @@ where
             }
 
             // Check if the connection (and handler) should be shut down.
-            // As long as we're still negotiating substreams or have any active streams shutdown is always postponed.
+            // As long as we're still negotiating substreams or have
+            // any active streams shutdown is always postponed.
             if negotiating_in.is_empty()
                 && negotiating_out.is_empty()
                 && requested_substreams.is_empty()
@@ -419,7 +428,9 @@ where
                             stream_counter.clone(),
                         ));
 
-                        continue; // Go back to the top, handler can potentially make progress again.
+                        // Go back to the top,
+                        // handler can potentially make progress again.
+                        continue;
                     }
                 }
             }
@@ -436,7 +447,9 @@ where
                             stream_counter.clone(),
                         ));
 
-                        continue; // Go back to the top, handler can potentially make progress again.
+                        // Go back to the top,
+                        // handler can potentially make progress again.
+                        continue;
                     }
                 }
             }
@@ -451,10 +464,12 @@ where
                 for change in changes {
                     handler.on_connection_event(ConnectionEvent::LocalProtocolsChange(change));
                 }
-                continue; // Go back to the top, handler can potentially make progress again.
+                // Go back to the top, handler can potentially make progress again.
+                continue;
             }
 
-            return Poll::Pending; // Nothing can make progress, return `Pending`.
+            // Nothing can make progress, return `Pending`.
+            return Poll::Pending;
         }
     }
 
@@ -482,7 +497,8 @@ fn compute_new_shutdown(
 ) -> Option<Shutdown> {
     match (current_shutdown, handler_keep_alive) {
         (_, false) if idle_timeout == Duration::ZERO => Some(Shutdown::Asap),
-        (Shutdown::Later(_), false) => None, // Do nothing, i.e. let the shutdown timer continue to tick.
+        // Do nothing, i.e. let the shutdown timer continue to tick.
+        (Shutdown::Later(_), false) => None,
         (_, false) => {
             let now = Instant::now();
             let safe_keep_alive = checked_add_fraction(now, idle_timeout);
@@ -493,10 +509,12 @@ fn compute_new_shutdown(
     }
 }
 
-/// Repeatedly halves and adds the [`Duration`] to the [`Instant`] until [`Instant::checked_add`] succeeds.
+/// Repeatedly halves and adds the [`Duration`]
+/// to the [`Instant`] until [`Instant::checked_add`] succeeds.
 ///
-/// [`Instant`] depends on the underlying platform and has a limit of which points in time it can represent.
-/// The [`Duration`] computed by the this function may not be the longest possible that we can add to `now` but it will work.
+/// [`Instant`] depends on the underlying platform and has a limit of which points in time it can
+/// represent. The [`Duration`] computed by the this function may not be the longest possible that
+/// we can add to `now` but it will work.
 fn checked_add_fraction(start: Instant, mut duration: Duration) -> Duration {
     while start.checked_add(duration).is_none() {
         tracing::debug!(start=?start, duration=?duration, "start + duration cannot be presented, halving duration");
@@ -516,7 +534,7 @@ pub(crate) struct IncomingInfo<'a> {
     pub(crate) send_back_addr: &'a Multiaddr,
 }
 
-impl<'a> IncomingInfo<'a> {
+impl IncomingInfo<'_> {
     /// Builds the [`ConnectedPoint`] corresponding to the incoming connection.
     pub(crate) fn create_connected_point(&self) -> ConnectedPoint {
         ConnectedPoint::Listener {
@@ -767,18 +785,22 @@ impl<T: AsRef<str>> std::hash::Hash for AsStrHashEq<T> {
 
 #[cfg(test)]
 mod tests {
+    use std::{
+        convert::Infallible,
+        sync::{Arc, Weak},
+        time::Instant,
+    };
+
+    use futures::{future, AsyncRead, AsyncWrite};
+    use libp2p_core::{
+        upgrade::{DeniedUpgrade, InboundUpgrade, OutboundUpgrade, UpgradeInfo},
+        StreamMuxer,
+    };
+    use quickcheck::*;
+    use tracing_subscriber::EnvFilter;
+
     use super::*;
     use crate::dummy;
-    use futures::future;
-    use futures::AsyncRead;
-    use futures::AsyncWrite;
-    use libp2p_core::upgrade::{DeniedUpgrade, InboundUpgrade, OutboundUpgrade, UpgradeInfo};
-    use libp2p_core::StreamMuxer;
-    use quickcheck::*;
-    use std::sync::{Arc, Weak};
-    use std::time::Instant;
-    use tracing_subscriber::EnvFilter;
-    use void::Void;
 
     #[test]
     fn max_negotiating_inbound_streams() {
@@ -906,7 +928,8 @@ mod tests {
         );
         assert!(connection.handler.remote_removed.is_empty());
 
-        // Third, stop listening on a protocol it never advertised (we can't control what handlers do so this needs to be handled gracefully).
+        // Third, stop listening on a protocol it never advertised (we can't control what handlers
+        // do so this needs to be handled gracefully).
         connection.handler.remote_removes_support_for(&["/baz"]);
         let _ = connection.poll_noop_waker();
 
@@ -1016,7 +1039,7 @@ mod tests {
 
     impl StreamMuxer for DummyStreamMuxer {
         type Substream = PendingSubstream;
-        type Error = Void;
+        type Error = Infallible;
 
         fn poll_inbound(
             self: Pin<&mut Self>,
@@ -1051,7 +1074,7 @@ mod tests {
 
     impl StreamMuxer for PendingStreamMuxer {
         type Substream = PendingSubstream;
-        type Error = Void;
+        type Error = Infallible;
 
         fn poll_inbound(
             self: Pin<&mut Self>,
@@ -1113,7 +1136,7 @@ mod tests {
 
     struct MockConnectionHandler {
         outbound_requested: bool,
-        error: Option<StreamUpgradeError<Void>>,
+        error: Option<StreamUpgradeError<Infallible>>,
         upgrade_timeout: Duration,
     }
 
@@ -1133,7 +1156,7 @@ mod tests {
 
     #[derive(Default)]
     struct ConfigurableProtocolConnectionHandler {
-        events: Vec<ConnectionHandlerEvent<DeniedUpgrade, (), Void>>,
+        events: Vec<ConnectionHandlerEvent<DeniedUpgrade, (), Infallible>>,
         active_protocols: HashSet<StreamProtocol>,
         local_added: Vec<Vec<StreamProtocol>>,
         local_removed: Vec<Vec<StreamProtocol>>,
@@ -1166,8 +1189,8 @@ mod tests {
     }
 
     impl ConnectionHandler for MockConnectionHandler {
-        type FromBehaviour = Void;
-        type ToBehaviour = Void;
+        type FromBehaviour = Infallible;
+        type ToBehaviour = Infallible;
         type InboundProtocol = DeniedUpgrade;
         type OutboundProtocol = DeniedUpgrade;
         type InboundOpenInfo = ();
@@ -1194,13 +1217,13 @@ mod tests {
                 ConnectionEvent::FullyNegotiatedInbound(FullyNegotiatedInbound {
                     protocol,
                     ..
-                }) => void::unreachable(protocol),
+                }) => libp2p_core::util::unreachable(protocol),
                 // TODO: remove when Rust 1.82 is MSRV
                 #[allow(unreachable_patterns)]
                 ConnectionEvent::FullyNegotiatedOutbound(FullyNegotiatedOutbound {
                     protocol,
                     ..
-                }) => void::unreachable(protocol),
+                }) => libp2p_core::util::unreachable(protocol),
                 ConnectionEvent::DialUpgradeError(DialUpgradeError { error, .. }) => {
                     self.error = Some(error)
                 }
@@ -1216,7 +1239,7 @@ mod tests {
         fn on_behaviour_event(&mut self, event: Self::FromBehaviour) {
             // TODO: remove when Rust 1.82 is MSRV
             #[allow(unreachable_patterns)]
-            void::unreachable(event)
+            libp2p_core::util::unreachable(event)
         }
 
         fn connection_keep_alive(&self) -> bool {
@@ -1246,8 +1269,8 @@ mod tests {
     }
 
     impl ConnectionHandler for ConfigurableProtocolConnectionHandler {
-        type FromBehaviour = Void;
-        type ToBehaviour = Void;
+        type FromBehaviour = Infallible;
+        type ToBehaviour = Infallible;
         type InboundProtocol = ManyProtocolsUpgrade;
         type OutboundProtocol = DeniedUpgrade;
         type InboundOpenInfo = ();
@@ -1293,7 +1316,7 @@ mod tests {
         fn on_behaviour_event(&mut self, event: Self::FromBehaviour) {
             // TODO: remove when Rust 1.82 is MSRV
             #[allow(unreachable_patterns)]
-            void::unreachable(event)
+            libp2p_core::util::unreachable(event)
         }
 
         fn connection_keep_alive(&self) -> bool {
@@ -1333,7 +1356,7 @@ mod tests {
 
     impl<C> InboundUpgrade<C> for ManyProtocolsUpgrade {
         type Output = C;
-        type Error = Void;
+        type Error = Infallible;
         type Future = future::Ready<Result<Self::Output, Self::Error>>;
 
         fn upgrade_inbound(self, stream: C, _: Self::Info) -> Self::Future {
@@ -1343,7 +1366,7 @@ mod tests {
 
     impl<C> OutboundUpgrade<C> for ManyProtocolsUpgrade {
         type Output = C;
-        type Error = Void;
+        type Error = Infallible;
         type Future = future::Ready<Result<Self::Output, Self::Error>>;
 
         fn upgrade_outbound(self, stream: C, _: Self::Info) -> Self::Future {